+++ server/prisma/migrations/20260428185433_add_iva_tasa/migration.sql
@@ -1,0 +1,11 @@
+-- AlterTable
+ALTER TABLE "finance_movements" ADD COLUMN     "ivaTasa" DOUBLE PRECISION DEFAULT 22.0;
+
+-- AlterTable
+ALTER TABLE "invoice_items" ADD COLUMN     "ivaTasa" DOUBLE PRECISION NOT NULL DEFAULT 22.0;
+
+-- AlterTable
+ALTER TABLE "material_items" ADD COLUMN     "ivaTasa" DOUBLE PRECISION NOT NULL DEFAULT 22.0;
+
+-- AlterTable
+ALTER TABLE "project_materials" ADD COLUMN     "ivaTasa" DOUBLE PRECISION NOT NULL DEFAULT 22.0;